ok im done now
